Move files in integration tests
