+++ integration-tests/assetft/global_freeze.go
@@ -1,4 +1,4 @@
-package asset
+package assetft
 
 import (
 	"context"
