+++ integration-tests/assetft/issue.go
@@ -1,4 +1,4 @@
-package asset
+package assetft
 
 import (
 	"context"
